+++ matchbox-engine/src/main/java/org/hl7/fhir/r5/context/BaseWorkerContext.java
@@ -3332,7 +3332,7 @@ public abstract class BaseWorkerContext extends I18nBase implements IWorkerConte
         return (T) scs.getCs();
       }
     } else {
-      throw new Error("Not supported");
+      throw new Error("Not supported: doFindTxResource with type of "+class_.getName());
     }
   }