+++ docs/changelog.md
@@ -1,6 +1,6 @@
 Unreleased
 
-- Upgrade hapifhir org.hl7.fhir.core to 6.4.4
+- Upgrade org.hl7.fhir.core to 6.4.4 and hapi-fhir to 7.6.0
 - Remove the `devMode` configuration parameter, it is now enabled when `httpReadOnly` is not [#315](https://github.com/ahdis/matchbox/issues/315)
 - Remove the `autoInstallMissingIgs` configuration parameter, it is now enabled when `httpReadOnly` is not [#315](https://github.com/ahdis/matchbox/issues/315)
 - Improve the Matchbox server documentation [#315](https://github.com/ahdis/matchbox/issues/315)